# swagger.json
{
	"openapi": "3.0.3",
	"info": {
		"title": "Meu Amigo PET",
		"version": "1.0.0",
		"description": "Meu Amigo PET é uma API REST cujo objetivo principal é conectar usuários, sejam eles pais de PETs ou aspirantes ao cargo, com objetivos semelhantes: adoção ou cruzamento. A conexão entre usuários se dá por meio de anúncios, os quais podem ser de adoção ou cruzamento. Assim, usuários logados podem acessar o 'feed de anúnicos' e obter informações de pets e usuários. A API visa atender usuários comuns e administradores com funcionalidades diferentes para cada um dos tipos de contas. Os usuários devem estar logados no servidor para acessar as diferentes rotas e suas funcionalidades. A aplicação foi construida usando-se Node.JS, módulos NPM (específicados no arquivo **package.json**), o framework Express e o [GitHub](https: //github.com/Alisson3366/projeto-de-back-end)."
	},
	"servers": [
		{
			"url": "http://localhost:8080"
		}
	],
...
	"paths": {
...
		"/registro": {
			"post": {
				"tags": ["inicio"],
				"summary": "Registro na API.",
				"description": "Registro na API.",
				"operationId": "registrar",
				"requestBody": {
					"description": "Registre-se na API informando email, senha e nome.",
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"email": {
										"type": "string",
										"example": "user1@mail.com"
									},
									"senha": {
										"type": "string",
										"example": "teste1"
									},
									"nome": {
										"type": "string",
										"example": "ARNALDO SANTIAGO"
									}
								}
							}
						}
					},
					"required": true
				},
				"responses": {
					"201": {
						"description": "Novo usuário registrado com sucesso!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Mensagem": {
											"type": "string"
										},
										"Usuario": {
											"type": "object"
										}
									}
								}
							}
						}
					},
					"422": {
						"description": "Nome, email ou senha inválidos!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Erro interno na aplicação!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					}
				}
			}
		},
...
		"/anuncios": {
			"get": {
				"tags": ["anuncios"],
				"summary": "Consulta de todos os anúncios",
				"description": "Consulta de todos os anúncios",
				"operationId": "consultaAnuncios",
				"responses": {
					"200": {
						"description": "Consulta efetuada com sucesso!",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/ANUNCIOS"
									}
								}
							}
						}
					},
					"500": {
						"description": "Erro interno na aplicação!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					}
				}
			}
		},
		"/anuncio/{id}": {
			"get": {
				"tags": ["anuncios"],
				"summary": "Consulta de anúncio específico pelo ID",
				"description": "Consulta de anúncio específico pelo ID",
				"operationId": "consultaAnuncioID",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "ID do anúncio a ser consultado.",
						"required": true,
						"schema": {
							"type": "string",
							"example": "3732f525-2632-4460-b67d-fcad2262e8f1"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Consulta efetuada com sucesso!",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ANUNCIOS"
								}
							}
						}
					},
					"400": {
						"description": "ID de anúncio inválido!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Anúncio não localizado!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Erro interno na aplicação!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					}
				}
			}
		},
		"/anuncios/usuario": {
			"get": {
				"tags": ["anuncios"],
				"summary": "Consulta dos anúncios do usuário logado",
				"description": "Consulta dos anúncios do usuário logado",
				"operationId": "consultaAnunciosUsuario",
				"responses": {
					"200": {
						"description": "Consulta efetuada com sucesso!",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/ANUNCIOS"
									}
								}
							}
						}
					},
					"401": {
						"description": "Token não informado ou inválido. Acesso negado!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Usuário não localizado!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Erro interno na aplicação!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					}
				}
			},
			"post": {
				"tags": ["anuncios"],
				"summary": "Cria novo anúncio do usuário logado",
				"description": "Cria novo anúncio do usuário logado",
				"operationId": "adicionaAnuncioUsuario",
				"requestBody": {
					"description": "Cria novo anúncio do usuário logado",
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"tipo": {
										"type": "string",
										"example": "cruzamento"
									},
									"titulo": {
										"type": "string",
										"example": "PROCURO UM PARCEIRO DE CRUZA PARA A MINHA BORDER COLLIE"
									},
									"sexo": {
										"type": "string",
										"example": "fêmea"
									},
									"raca": {
										"type": "string",
										"example": "border collie"
									},
									"quantidade": {
										"type": "integer",
										"example": 0
									}
								}
							}
						}
					},
					"required": true
				},
				"responses": {
					"201": {
						"description": "Anúncio adicionado com sucesso!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Mensagem": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Token não informado ou inválido. Acesso negado!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Usuário não localizado!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"422": {
						"description": "Tipo, titulo, sexo, raca ou quantidade inválidos!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Erro interno na aplicação!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					}
				}
			}
		},
		"/anuncio/usuario/{id}": {
			"patch": {
				"tags": ["anuncios"],
				"summary": "Atualiza um anúncio já existente de um usuário",
				"description": "Atualiza um anúncio já existente de um usuário",
				"operationId": "atualizaAnuncioUsuario",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "ID do anúncio a ser atualizado.",
						"required": true,
						"schema": {
							"type": "string",
							"example": "3732f525-2632-4460-b67d-fcad2262e8f1"
						}
					}
				],
				"requestBody": {
					"description": "Atualiza um anúncio já existente de um usuário",
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"tipo": {
										"type": "string",
										"example": "cruzamento"
									},
									"titulo": {
										"type": "string",
										"example": "PROCURO UM PARCEIRO DE CRUZA PARA A MINHA BORDER COLLIE"
									},
									"sexo": {
										"type": "string",
										"example": "fêmea"
									},
									"raca": {
										"type": "string",
										"example": "border collie"
									},
									"quantidade": {
										"type": "integer",
										"example": 0
									}
								}
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Anúncio atualizado com sucesso!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Mensagem": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "ID de anúncio inválido!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Token não informado ou inválido. Acesso negado!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Anúncio não localizado!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"422": {
						"description": "Tipo, titulo, sexo, raca ou quantidade inválidos!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Erro interno na aplicação!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					}
				}
			},
			"delete": {
				"tags": ["anuncios"],
				"summary": "Deleta o anúncio de um usuário da API",
				"description": "Deleta o anúncio de um usuário da API",
				"operationId": "deletaAnuncioUsuario",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "ID do anúncio a ser deletado.",
						"required": true,
						"schema": {
							"type": "string",
							"example": "3732f525-2632-4460-b67d-fcad2262e8f1"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Anúncio deletado com sucesso!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Mensagem": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "ID de anúncio inválido!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Token não informado ou inválido. Acesso negado!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Anúncio não localizado!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Erro interno na aplicação!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					}
				}
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"ANUNCIOS": {
				"required": ["tipo", "titulo", "sexo", "raca"],
				"type": "object",
				"properties": {
					"_id": {
						"type": "string",
						"example": "3732f525-2632-4460-b67d-fcad2262e8f1"
					},
					"donoAnuncio": {
						"type": "string",
						"example": "3732f525-2632-4460-b67d-fcad2262e8f1"
					},
					"tipo": {
						"type": "string",
						"example": "adoção"
					},
					"titulo": {
						"type": "string",
						"example": "3 FILHOTES DE PIT BUL PARA ADOÇÃO"
					},
					"sexo": {
						"type": "string",
						"example": "ambos"
					},
					"raca": {
						"type": "string",
						"example": "american pit bull terrier"
					},
					"quantidade": {
						"type": "integer",
						"example": 4
					}
				}
			},
...
		}
	}
}